tech(TTB-060): Moved Card components into individual directory
